requirements.txt
# Core Framework (versões flexíveis para compatibilidade com Agent Framework)
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.11.3  # Requerido pelo agent-framework
python-dotenv>=1.0.0

# Microsoft Agent Framework (substitui AutoGen)
agent-framework>=1.0.0b251016

# Data Processing (versões flexíveis para compatibilidade)
pandas>=2.1.3
numpy>=1.26.0

# Utilities
python-multipart>=0.0.6
aiofiles>=24.1.0  # Requerido pelo agent-framework
